+++ src/block/icon-list-old/editor.scss
