+++ src/block/icon-list-old/index.js
@@ -12,7 +12,6 @@ import metadata from './block.json'
 import schema from './schema'
 import example from './example'
 import deprecated from './deprecated'
-import transforms from './transforms'
 
 /**
  * WordPress dependencies
@@ -26,13 +25,18 @@ export const settings = {
 	supports: {
 		anchor: true,
 		spacing: true,
-		__unstablePasteTextInline: true,
-		__experimentalSelector: 'ol,ul',
-		__experimentalOnMerge: true,
 	},
 	example,
 	deprecated,
 	edit,
 	save,
-	transforms,
+	merge( attributes, attributesToMerge ) {
+		// Make sure that the selection is always at the end of the text.
+		// @see https://github.com/WordPress/gutenberg/blob/3da717b8d0ac7d7821fc6d0475695ccf3ae2829f/packages/block-library/src/paragraph/index.js
+		return {
+			text:
+				( attributes.text || '' ) +
+				( attributesToMerge.text || '' ),
+		}
+	},
 }
