+++ src/block/icon-list-old/style.scss
@@ -1,20 +1,20 @@
 // Important, we can't use flex since it causes issues with columns.
-.stk-block-icon-list.stk-block-icon-list-new {
+.stk-block-icon-list {
 	:is(ul, ol) {
-		width: fit-content; // TODO: Confirm if will set to 100% or not
+		width: fit-content;
 		text-align: start;
 	}
 	ol {
 		list-style-image: none;
 	}
 
-	ul.stk-block-icon-list__ul {
-		list-style-type: none;
-		padding-inline-start: 0px;
+	&.stk-block :is(ul, ol) {
+		padding-inline-start: 1em;
 	}
 
-	&.stk-block :is(ol) {
-		padding-inline-start: 1em;
+	// Line height creates unnecessary row gaps.
+	::marker {
+		line-height: 0;
 	}
 
 	// @from https://css-tricks.com/almanac/properties/b/break-inside/
@@ -22,16 +22,22 @@
 		page-break-inside: avoid; /* Firefox */
 		break-inside: avoid;
 	}
+
+	// Need to make this fit the content so that we can align left/center/right
+	// the list alignment.
+	li {
+		width: fit-content;
+	}
 }
 
 // Alignments when the icon list is aligned.
-.stk-block.stk-block-icon-list-new :is(ul.stk-block-icon-list__ul, ol.stk-block-icon-list__ol) {
+.stk-block-icon-list :is(ul, ol) {
 	margin-left: var(--stk-alignment-margin-left);
 	margin-right: var(--stk-alignment-margin-right);
 }
 
 // Prevents theme's css from affecting the icon list.
-.stk-block.stk-block-icon-list-new :is(ul.stk-block-icon-list__ul, ol.stk-block-icon-list__ol) {
+.stk-block-icon-list :is(ul, ol) {
 	margin-top: 0;
 	margin-bottom: 0;
