+++ src/block/icon-list/block.json
@@ -10,6 +10,11 @@
 		"Bullets",
 		"Number list"
 	],
+	"providesContext": {
+		"stackable/ordered": "ordered",
+		"stackable/uniqueId": "uniqueId",
+		"stackable/isIndented": "isIndented"
+	},
 	"textdomain": "stackable-ultimate-gutenberg-blocks",
 	"stk-type": "essential",
 	"stk-demo": "https://wpstackable.com/icon-list-block/?utm_source=welcome&utm_medium=settings&utm_campaign=view_demo&utm_content=demolink"
