+++ src/block/icon-list/edit.js
@@ -15,7 +15,7 @@ import {
 	AdvancedRangeControl,
 	IconControl,
 	ColorPaletteControl,
-	IconSearchPopover,
+	// IconSearchPopover,
 	AdvancedSelectControl,
 	AlignButtonsControl,
 } from '~stackable/components'
@@ -38,17 +38,24 @@ import {
 	getAlignmentClasses,
 	Transform,
 } from '~stackable/block-components'
+import { useBlockContext } from '~stackable/hooks'
 
 /**
  * WordPress dependencies
  */
-import { useSelect } from '@wordpress/data'
-import {
-	useRef, useEffect, useState,
-} from '@wordpress/element'
 import { sprintf, __ } from '@wordpress/i18n'
-import { createIconListControls, DEFAULT_SVG } from './util'
+import { DEFAULT_SVG, IconSvgDef } from './util'
 import { compose } from '@wordpress/compose'
+import { useInnerBlocksProps } from '@wordpress/block-editor'
+import { dispatch } from '@wordpress/data'
+import { useEffect } from '@wordpress/element'
+import { addFilter } from '@wordpress/hooks'
+
+const ALLOWED_INNER_BLOCKS = [ 'stackable/icon-list-item' ]
+
+const TEMPLATE = [
+	[ 'stackable/icon-list-item', {	text: '' } ],
+]
 
 const listTypeOptions = [
 	{
@@ -89,93 +96,31 @@ const listStyleTypeOptions = [
 ]
 
 const Edit = props => {
-	const textRef = useRef()
-	const typographyWrapperRef = useRef()
-	const [ isOpenIconSearch, setIsOpenIconSearch ] = useState( false )
-	const [ iconSearchAnchor, setIconSearchAnchor ] = useState( null )
-	const [ selectedIconCSSSelector, setSelectedIconCSSSelector ] = useState( null )
-	const [ selectedEvent, setSelectedEvent ] = useState( null )
-	const {
-		isTyping,
-	} = useSelect( select => ( {
-		isTyping: select( 'core/block-editor' ).isTyping(),
-	} ) )
-
-	useEffect( () => {
-		// Click handler to detect whether an icon is clicked, and open the icon
-		// picker for that icon.
-		const iconClickHandler = event => {
-			// If li isn't clicked, close the icon search.
-			setSelectedEvent( event )
-			if ( event.target.tagName !== 'LI' || event.target.parentElement.tagName !== 'UL' ) {
-				return setIsOpenIconSearch( false )
-			}
-
-			/**
-			 * Check if the click is on the icon.
-			 */
-
-			if ( event.offsetX <= 2 ) {
-				// Get the selected li and show the icon picker on it.
-				const index = Array.from( event.target.parentElement.children ).indexOf( event.target ) + 1
-				const { currentlyOpenIndex } = event.target.parentElement
-
-				if ( currentlyOpenIndex && currentlyOpenIndex === index ) {
-					event.target.parentElement.currentlyOpenIndex = undefined
-					return setIsOpenIconSearch( false )
-				}
-
-				event.target.parentElement.currentlyOpenIndex = index
-
-				/**
-				 * Get the CSS selector of the selected icon.
-				 *
-				 * @since 3.0.0
-				 */
-				let traverseToRichText = event.target
-				let found = null
-				while ( traverseToRichText.tagName !== 'DIV' ) {
-					for ( let i = 0; i < Array.from( traverseToRichText.parentElement.childNodes ).length; i++ ) {
-						if ( traverseToRichText.parentElement.childNodes[ i ] === traverseToRichText && found === null ) {
-							found = i + 1
-							break
-						}
-					}
-					traverseToRichText = traverseToRichText.parentElement
-				}
-
-				setSelectedIconCSSSelector( `ul li:nth-child(${ found })` )
-				setIconSearchAnchor( event.target )
-				return setIsOpenIconSearch( true )
-			}
-
-			// Hide the icon picker.
-			event.target.parentElement.currentlyOpenIndex = undefined
-			setIconSearchAnchor( null )
-			return setIsOpenIconSearch( false )
-		}
-
-		textRef.current.addEventListener( 'click', iconClickHandler )
-		return () => textRef.current?.removeEventListener( 'click', iconClickHandler )
-	}, [ setSelectedEvent, setSelectedIconCSSSelector, setIconSearchAnchor, setIsOpenIconSearch, textRef.current ] )
-
 	const {
 		clientId,
 		attributes,
 		setAttributes,
 		className,
 		isSelected,
-		onRemove,
-		mergeBlocks,
 	} = props
 
 	useGeneratedCss( props.attributes )
 
-	const { ordered } = attributes
-	const tagName = ordered ? 'ol' : 'ul'
+	const { ordered, icon } = attributes
+	const TagName = ordered ? 'ol' : 'ul'
+	const tagNameClass = ordered ? 'stk-block-icon-list__ol' : 'stk-block-icon-list__ul'
 
 	const textClasses = getTypographyClasses( attributes )
 	const blockAlignmentClass = getAlignmentClasses( attributes )
+	const { innerBlocks, parentBlock } = useBlockContext()
+
+	useEffect( () => {
+		if ( parentBlock && parentBlock.name === 'stackable/icon-list-item' ) {
+			setAttributes( { isIndented: true } )
+		} else {
+		 setAttributes( { isIndented: false } )
+		}
+	}, [ parentBlock ] )
 
 	const blockClassNames = classnames( [
 		className,
@@ -184,8 +129,20 @@ const Edit = props => {
 		textClasses,
 	] )
 
-	const controls = createIconListControls( {
-		isSelected, tagName,
+	const resetCustomIcons = () => {
+		innerBlocks.forEach( block => {
+			dispatch( 'core/block-editor' ).updateBlockAttributes( block.clientId, { icon: '' } )
+		} )
+	}
+
+	const innerBlocksProps = useInnerBlocksProps( {
+		className: tagNameClass,
+	}, {
+		allowedBlocks: ALLOWED_INNER_BLOCKS,
+		template: TEMPLATE,
+		templateInsertUpdatesSelection: true,
+		renderAppender: false,
+		__experimentalCaptureToolbars: true,
 	} )
 
 	return (
@@ -268,8 +225,9 @@ const Edit = props => {
 								label={ __( 'Icon', i18n ) }
 								value={ attributes.icon }
 								onChange={ icon => {
+									setAttributes( { icon } )
 									// Reset custom individual icons.
-									setAttributes( { icon, icons: [] } )
+									resetCustomIcons()
 								} }
 								defaultValue={ DEFAULT_SVG }
 							/>
@@ -352,59 +310,8 @@ const Edit = props => {
 				attributes={ props.attributes }
 				className={ blockClassNames }
 			>
-				{ /* eslint-disable-next-line */ }
-				<div
-					ref={ typographyWrapperRef }
-					onClick={ e => {
-						// When the div wrapper is clicked, pass the focus to the list.
-						if ( e.target === typographyWrapperRef.current ) {
-							textRef.current.focus()
-							const range = document.createRange()
-							range.selectNodeContents( textRef.current )
-							range.collapse( false )
-
-							const selection = window.getSelection() // eslint-disable-line
-							selection.removeAllRanges()
-							selection.addRange( range )
-						}
-					} }
-				>
-					<Typography
-						tagName={ tagName }
-						multiline="li"
-						onRemove={ onRemove }
-						onMerge={ mergeBlocks }
-						ref={ textRef }
-					>
-						{ controls }
-					</Typography>
-					{ ! isTyping && isSelected && isOpenIconSearch && (
-						<IconSearchPopover
-							__deprecatedAnchorRef={ iconSearchAnchor }
-							__deprecatedPosition="bottom left"
-							__hasPopover={ true }
-							onClose={ () => {
-								if ( selectedEvent ) {
-									selectedEvent.target.parentElement.currentlyOpenIndex = undefined
-								}
-								setIsOpenIconSearch( false )
-							} }
-							onChange={ icon => {
-								const icons = { ...props.attributes.icons }
-								if ( ! icon ) {
-									// Remove the icon inside the icons.
-									delete icons[ selectedIconCSSSelector ]
-								} else {
-									icons[ selectedIconCSSSelector ] = icon
-								}
-
-								setAttributes( {
-									icons: { ...icons },
-								} )
-							} }
-						/>
-					) }
-				</div>
+				{ ! ordered && <IconSvgDef icon={ icon } uniqueId={ attributes.uniqueId } /> }
+				<TagName { ...innerBlocksProps } />
 			</BlockDiv>
 			{ props.isHovered && <MarginBottom /> }
 		</>
@@ -416,3 +323,7 @@ export default compose(
 	withQueryLoopContext,
 	withBlockAttributeContext,
 )( Edit )
+
+addFilter( 'stackable.edit.margin-bottom.enable-handlers', 'stackable/icon-list', ( enabled, parentBlock ) => {
+	return parentBlock?.name === 'stackable/icon-list-item' ? false : enabled
+} )
