+++ src/block/icon-list/schema.js
@@ -52,10 +52,6 @@ export const iconListAttributes = {
 		default: '',
 		stkHover: true,
 	},
-	icons: {
-		type: 'object',
-		default: {},
-	},
 	iconSize: {
 		stkResponsive: true,
 		type: 'number',
@@ -86,6 +82,12 @@ export const iconListAttributes = {
 		type: 'number',
 		default: '',
 	},
+
+	// Indentation.
+	isIdented: {
+		type: 'boolean',
+		default: false,
+	},
 }
 
 export const attributes = ( version = VERSION ) => {
@@ -103,8 +105,6 @@ export const attributes = ( version = VERSION ) => {
 	ConditionalDisplay.addAttributes( attrObject )
 	Typography.addAttributes( attrObject, 'ul,ol', {
 		hasTextTag: false,
-		multiline: 'li',
-		multilineWrapperTags: [ 'ol', 'ul' ],
 	} )
 	MarginBottom.addAttributes( attrObject )
