+++ src/block/icon-list/style.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { convertSVGStringToBase64 } from './util'
+// import { convertSVGStringToBase64 } from './util'
 
 /**
  * External dependencies
@@ -35,13 +35,10 @@ const Styles = props => {
 		versionAdded: '3.0.0',
 		versionDeprecated: '',
 	}
-	const {
-		icons,
-	} = props
-
+	// TODO: clean up comments
 	return (
 		<>
-			{ Object.keys( icons ).reduce( ( acc, key ) => {
+			{ /* { Object.keys( icons ).reduce( ( acc, key ) => {
 				acc.push(
 					<BlockCss
 						{ ...propsToPass }
@@ -92,14 +89,23 @@ const Styles = props => {
 					/>
 				)
 				return acc
-			}, [] ) }
+			}, [] ) } */ }
 
 			<BlockCss
 				{ ...propsToPass }
-				selector="li"
+				selector="ol li"
 				styleRule="paddingInlineStart"
 				attrName="iconGap"
-				key="iconGap"
+				key="olIconGap"
+				responsive="all"
+				format="%spx"
+			/>
+			<BlockCss
+				{ ...propsToPass }
+				selector="ul li"
+				styleRule="gap"
+				attrName="iconGap"
+				key="ulIconGap"
 				responsive="all"
 				format="%spx"
 			/>
@@ -145,7 +151,7 @@ const Styles = props => {
 				responsive="all"
 				format="%spx"
 			/>
-			<BlockCss
+			{ /* <BlockCss
 				{ ...propsToPass }
 				selector="ul li"
 				hover="all"
@@ -171,8 +177,8 @@ const Styles = props => {
 					return `url('data:image/svg+xml;base64,${ iconWithColor }')`
 				} }
 				dependencies={ [ 'icon', 'iconRotation', 'iconOpacity' ] }
-			/>
-			<BlockCss
+			/> */ }
+			{ /* <BlockCss
 				{ ...propsToPass }
 				selector="li::marker"
 				hover="all"
@@ -189,7 +195,7 @@ const Styles = props => {
 				key="iconSize"
 				responsive="all"
 				format="%sem"
-			/>
+			/> */ }
 			<BlockCss
 				{ ...propsToPass }
 				// For calculating the approximate clickable area for
@@ -250,7 +256,7 @@ IconListStyles.Content = props => {
 			<EffectsAnimations.Style.Content { ...props } />
 			<Advanced.Style.Content { ...props } />
 			<Transform.Style.Content { ...props } />
-			<Styles { ...props } icons={ props.attributes.icons } />
+			<Styles { ...props } />
 		</BlockCssCompiler>
 	)
 }
